Set AzureWrapper as public dependency and Move Internal header to public folder
