+++ Source/AzSpeech/Private/AzSpeech/AzSpeechHelper.cpp
@@ -3,7 +3,7 @@
 // Repo: https://github.com/lucoiso/UEAzSpeech
 
 #include "AzSpeech/AzSpeechHelper.h"
-#include "AzSpeech/AzSpeechInternalFuncs.h"
+#include "AzSpeechInternalFuncs.h"
 #include <Sound/SoundWave.h>
 #include <Misc/FileHelper.h>
 #include <Misc/Paths.h>
