+++ Source/AzSpeech/Private/AzSpeech/Tasks/RecognitionMapCheckAsync.cpp
@@ -4,8 +4,8 @@
 
 #include "AzSpeech/Tasks/RecognitionMapCheckAsync.h"
 #include "AzSpeech/AzSpeechRecognitionMap.h"
-#include "AzSpeech/AzSpeechInternalFuncs.h"
 #include "AzSpeech/AzSpeechSettings.h"
+#include "AzSpeechInternalFuncs.h"
 #include <Async/Async.h>
 
 URecognitionMapCheckAsync* URecognitionMapCheckAsync::RecognitionMapCheckAsync(const UObject* WorldContextObject, const FString& InString, const FName GroupName, const bool bStopAtFirstTrigger)
