+++ Source/AzSpeech/Public/AzSpeechInternalFuncs.h
